Refactor OMPMAPunchForm and OMPMAPunchFormContainer to use components from the new structure, removed unused files for OpplysningerOmOMPMASoknad and OMPMASoknadKvittering. Byttet alle less til tailwind for OMPMA
