--- src/app/søknader/omsorgspenger-midlertidig-alene/containers/OpplysningerOmSoknad/opplysningerOmOMPMASoknad.less
@@ -1,9 +1,0 @@
-.navds-alert {
-    margin-top: 1rem;
-    .navds-body-long {
-        font-size: 1rem !important;
-    }
-}
-.klokkeslett {
-    margin-left: 1rem;
-}
